mock/consumer/requirements.txt
kafka-python==2.0.2
pkg-resources==0.0.0
pprint==0.1
protobuf==3.14.0
six==1.15.0
